+++ flutter-login/example/lib/main.dart
@@ -31,6 +31,8 @@ class MyHomePage extends StatefulWidget {
 
 class _MyHomePageState extends State<MyHomePage> {
   Map<String, dynamic> _profile = {};
+
+  // Optionally, specify different "callbackUri" and "webCallbackUri" for Flutter Web.
   Auth3Login auth3 = Auth3Login(clientId: AUTH3_CLIENT_ID);
 
   void initState() {